Day Part 2 : solution fonctionnelle.

+++ src/main/2022/Day11/Simulation/Modele/Lancer.java
@@ -1,4 +1,4 @@
 package Day11.Simulation.Modele;
 
-public record Lancer(int singeDestinataire, int item) {
+public record Lancer(int singeDestinataire, Long item) {
 }
