+++ src/main/2022/Day11/Simulation/Modele/Operande/Operande.java
@@ -1,5 +1,5 @@
 package Day11.Simulation.Modele.Operande;
 
 public interface Operande {
-    Integer getValeur(Integer item);
+    Long getValeur(Long item);
 }
