+++ src/main/2022/Day11/Simulation/Modele/Operande/OperandeSelf.java
@@ -3,7 +3,7 @@ package Day11.Simulation.Modele.Operande;
 public class OperandeSelf implements Operande {
 
     @Override
-    public Integer getValeur(Integer item) {
+    public Long getValeur(Long item) {
         return item;
     }
 }
